Overhaul contour_functions to use opencv module
Deprecated matplotlib._cntr.Cntr replaced with
contour_functions.Contours class to interface with opencv (cv2) module

Identified contours using cv2.findContours() are defined by exact points
on the sampling grid, and as a result have a jagged appearance. The
previous Cntr.trace() function had some smoothing built in. Because the
old and new contour paths are not identical, the resulting identified
weighted centers (e.g., with the const area center and momentum deficit
trackers) are not identical to the old approach; in most cases, the
differences are O(0.1) or less, based on tracker_comparison.ipynb