Rewrite of optimziation algorithm
Motivated by sensitivity to initial guess, and erronenous convergence in
approximately uniform regions (resulting in incorrectly identified
wakes).
- First, an (easier) 1-D Gaussian problem is solved, with initial guess
being the center of the plane
- Second, the 2-D Gaussian problem is solved, with initial guess for yc
and zc being the solution to the 1-D problem.
- If the first step fails, then use location of largest velocity deficit
as the fallback guess.
- In the second step, solve a weighted least squares problem, with
weighting function identical to the exponential function in our Gaussian
definition. Hopefully, this decreases sensitivity to turbulent
structures in the background flow.
- Revert to approximate Jacobian (2-pt estimate, by default) because it
appears to be more robust. Also, the exact Jacobian has a tendency to
produce only circular wakes; further debugging will be required.