Fix uniform_filter; add tol option; update verbosity
- Previously filter was being applied over all array dimensions, i.e, in
time and in space.
- 'tol' kwarg is applied to ftol, xtol, and gtol (i.e., the change in
the cost function, change in independent variables, and the norm of the
gradient, respectively)
- Default tolerances increased to 1e-8 (to match least_squares default)

IMPORTANT NOTES:
- Using the uniform_filter with larger filter sizes (the default is D/5)
can artificially increase the size of your detected wake. This happens
because a larger filter will smooth out the field and reduce velocity
deficit maxima; as a result, the amplitude of the Gaussian will be lower
and it will have to expand to in size (sigma_y, sigma_z) to reduce
error.
- Decreasing 'tol' (e.g., to 1e-14) does not appear to appreciably
change the solution but degrades performance.